Add files via upload
The overall goal of this script was to create a machine learning model in tensorflow in order to analyze gut microbiome data as a means of predicting whether an individual has Type 1 diabetes based upon ASV counts of 256 microbes among ~16k samples. This model could theoretically be used as a means of preventative testing for Type 1 diabetes if tested against multiple datasets.